Bug fix and removed catch from main()
It's better to leave it uncaught here, since even after catching all
we do is print e.what() and exit.
Might as well let it call std::terminate.
Added benefit of leaving it uncaught is it causes a trap in a debugger.
Instead of just exiting in case of an exception.